Added rectangles with corners and circles

+++ OpgenGlTest1/ShapeInstance.h
@@ -18,6 +18,6 @@ struct ShapeInstance
 	Color color;
 
 	float radius;
-	int type;
+	float type;
 	float rotation;
 };
